requirements.txt
django==3.2
mysqlclient
Pillow
djangorestframework
django-cors-headers
cloudinary
djangorestframework-simplejwt
whitenoise